+++ Clouds.h
@@ -7,8 +7,8 @@ class Clouds : public Animation
 public:
 
 
-	Clouds(LPCTSTR bitmapPath, Graphics* graphics, float x = SCREEN_WIDTH / SCREEN_WIDTH - 1, 
-		float y = (int)(SCREEN_HEIGHT * 0.10), float xSpeed = 0.02);
+	Clouds(LPCTSTR bitmapPath, Graphics* graphics, int x = SCREEN_WIDTH / SCREEN_WIDTH - 1, 
+		int y = (SCREEN_HEIGHT * 0.10), float xSpeed = 0.02);
 	
 
 private:
